move types and enum to constants file
