+++ src/types/statsig.ts
@@ -1,0 +1,5 @@
+export type StatsigConfigType = Record<StatSigFlags, boolean>;
+
+export enum StatSigFlags {
+  ffSkipMigration = 'ff_skip_migration',
+}
